Все DWORD'ы, WORD'ы, BOOL'ы и прочие были заменены на стандартные названия типов данных во всём коде (Fixes issue #11). Makefile'ы улучшаются по мере того, как я всё больше погружаюсь в info make. Компиляция библиотеки движка в GCC теперь не требует флага совместимости -fpermissive, следовательно он был убран.
Велась работа над фиксом некоторых варнингов, а именно:
1. char* заменён на const char* там, где функция не изменяет его значение (не во всех местах, а только там где был варнинг, так что работу над этим 
надо продолжить).
2. Кое-какие объявления неиспользуемых переменных были закомментированы с пометкой unused variable.
3. В некоторых функциях, возвращающих значение, поставлен забытый return.
4. Порядок инициализации членов в конструкторах в некоторых местах отличался от порядка их описания в классе - поправил.
5. Кое-какие директивы препроцессора были заключены в _MSC_VER ifdef, дабы не генерировали варнингов о неподдерживаемости на gcc.

+++ Engine/inc/2de_GraphicsHigh.h
@@ -143,7 +143,7 @@ public:
 	RGBAf		color;					// цвет
 	GLuint		m_textureID;			// опенгловская хрень идентиф. текстуру
 
-	DWORD		ellapsedtime, lasttime;	// херь чтобы делать время между кадрами
+	unsigned long	ellapsedtime, lasttime;	// херь чтобы делать время между кадрами
 	int			m_nFrameNumber,			// номер текущего кадра
 		m_nCurrentRow,			// номер текущей строки
 		m_nCurrentColumn;		// номер текущей клонки
